+++ templatetags/templatetags/dict.py
@@ -5,4 +5,7 @@ register = template.Library()
 
 @register.filter
 def keyvalue(data, key):
-    return data[key]
+    if hasattr(data, 'get'):
+        return data.get(key)
+    else:
+        return data[key]