+++ var_stats_qc/__init__.py
@@ -1,0 +1,5 @@
+"""This module contains code for computing variant statistics and pedigree checks.
+
+It provides similar functionality to Peddy but works with a database backend. Peddy is described in
+Pedersen & Quinlan (2017).
+"""