+++ var_stats_qc/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class VarStatsQcConfig(AppConfig):
+    name = "var_stats_qc"