+++ variants/tests/_fixtures.py
@@ -1,9 +1,14 @@
 """Common fixture code.
 """
 
+import aldjemy.core
+
 from projectroles.models import Project
 from variants.models import SmallVariant
+from variants.variant_stats import rebuild_case_variant_stats
 
+#: The SQL Alchemy engine to use
+SQLALCHEMY_ENGINE = aldjemy.core.get_engine()
 
 #: Shared data for ``Project`` to use for all test cases.
 PROJECT_DICT = {
@@ -226,3 +231,5 @@ def fixture_setup_case1_simple():
         ensembl_hgvs_p="p.=",
         ensembl_effect=["synonymous_variant"],
     )
+
+    rebuild_case_variant_stats(SQLALCHEMY_ENGINE.connect(), case)